feat: home 앱 뱃지 모델 추가, challenge 앱 뱃지 모델 삭제
