+++ Dherethon/challenges/admin.py
@@ -6,4 +6,3 @@ admin.site.register(Challenge)
 admin.site.register(Goal)
 admin.site.register(UserChallenge)
 admin.site.register(GoalProgress)
-admin.site.register(Badge)
