+++ Dherethon/challenges/models.py
@@ -83,15 +83,3 @@ class GoalProgress(models.Model):
 
     def __str__(self):
         return f"{self.user.nickname} - {self.goal.content} - {'완료' if self.is_completed else '미완료'}"
-
-# 뱃지 부여
-class Badge(models.Model):
-    user = models.ForeignKey("api.User", on_delete=models.CASCADE)
-    category = models.ForeignKey(Category, on_delete=models.CASCADE, default=1) # default=1 임시 설정, 추후 수정 필요
-    challenge = models.ForeignKey(Challenge, on_delete=models.CASCADE)
-    awarded_at = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return f"{self.user.nickname} - {self.challenge.title}"
-
-
